% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bench-format.R
\name{bench_format}
\alias{bench_format}
\title{Function to format benchmarking results}
\usage{
bench_format(bench_res)
}
\arguments{
\item{bench_res}{benchmarking results}
}
\value{
formatted benchmarking results
}
\description{
Function to format benchmarking results
}
