% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-benchmark.R
\name{bench_sumplot}
\alias{bench_sumplot}
\title{Function that provides summary and plots for the benchmark run}
\usage{
bench_sumplot(.res_tible, title = "")
}
\arguments{
\item{.res_tible}{formatted bench result tible with added roc column}

\item{title}{character string for title of plots}
}
\value{
AUROC summary per row with TF coverage, ROC AUROC, Heatmap plots
}
\description{
Function that provides summary and plots for the benchmark run
}
