% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_tibble.R
\docType{data}
\name{bench_tibble}
\alias{bench_tibble}
\title{Benchmark input tibble containing the (experimental) design for each of the
benchmark runs corresponding to rows.}
\format{
A tibble with locations, options, and filter options for the desired benchmark setting

\describe{
\item{row_name}{User-defined name of row/run - e.g. network + benchmark data}
\item{net_loc}{Location of the network/gene set resource}
\item{lvls}{Column to filter by - e.g. confidence levels of the network}
\item{gene_source}{Source of nodes for the network AKA .source}
\item{target}{Network nodes AKA .target}
\item{statistics}{List of statistics to run}
\item{bnch_expr}{Location of the benchmark data set. The data set is in the form of a matrix
with each experiment summarized in a column}
\item{bench_meta}{Location of benchmark metadata. The metadata contains information about the
different experiments - e.g. type of method, cell type, etc.}
\item{opts}{Named list to be used to pass the options for each statistical method}
}
}
\description{
To be used as default input or as a starting point and
modified according to the benchmark run in mind.
}
\details{
Read and executed in a row-wise manner.
}
