% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-bench_res.R
\docType{class}
\name{BenchResult-class}
\alias{BenchResult-class}
\title{S4 Class used to format benchmark output.}
\description{
S4 Class used to format benchmark output.
}
\section{Slots}{

\describe{
\item{\code{bench_res}}{Formatted or non-formatted Benchmark output}

\item{\code{summary}}{Summary returned by the bench_sumplot functions - it countains
auroc and coverage summary table, roc plot, and auroc heatmap and barplot}

\item{\code{design}}{The input design tibble used to generate the benchmark results}
}}

