% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench-auroc.R
\name{calc_curve}
\alias{calc_curve}
\title{This function calculates precision recall curve}
\usage{
calc_curve(df, downsampling = T, times = 1000, ranked = F, curve = "PR")
}
\arguments{
\item{df}{run_benchmark roc column provided as input}

\item{downsampling}{logical flag indicating if the number of TN should be
downsampled to the number of TP}

\item{times}{integer showing the number of downsampling}

\item{ranked}{logical flag indicating if input is derived from composite
ranking that already took up-/downregulation (sign) into account}
}
\value{
tidy data frame with precision, recall, auc, n, tp, tn and coverage
}
\description{
This function calculates precision recall curve
}
