% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bench-auroc.R
\name{calc_roc_curve}
\alias{calc_roc_curve}
\title{This function calculates receiver operating characteristic}
\usage{
calc_roc_curve(df, downsampling = F, times = 1000, ranked = F)
}
\arguments{
\item{df}{run_benchmark roc column provided as input}

\item{downsampling}{logical flag indicating if the number of TN should be
downsampled to the number of TP}

\item{times}{integer showing the number of downsampling}

\item{ranked}{logical flag indicating if input is derived from composite
ranking that already took up-/downregulation (sign) into account}
}
\value{
tidy data frame with tpr, fpr, auc, n, tp, tn and coverage
}
\description{
This function calculates receiver operating characteristic
}
