% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var.R
\name{cast_as}
\alias{cast_as}
\title{Helper function to cast data from keyword of function}
\usage{
cast_as(variable, cast)
}
\arguments{
\item{variable}{String to be casted}

\item{cast}{Casting function or keyword}
}
\value{
casted variable
}
\description{
Helper function to cast data from keyword of function
}
