% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench-run-prep.R
\name{check_preced}
\alias{check_preced}
\title{Helper Function that checks if the preceding vector element is the same
as the current element}
\usage{
check_preced(vector_loc)
}
\arguments{
\item{vector_loc}{character vector with directory paths}
}
\value{
logical values describing whether the location of the loaded files
has changes
}
\description{
Helper Function that checks if the preceding vector element is the same
as the current element
}
\keyword{internal}
