% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench-run-prep.R
\name{check_prereq}
\alias{check_prereq}
\title{Helper Function that checks if the source_set/network set is appropriately
formatted}
\usage{
check_prereq(source_loc, target_col, source_col, filter_col)
}
\arguments{
\item{source_loc}{location of the source_set}
}
\description{
Helper Function that checks if the source_set/network set is appropriately
formatted
}
