% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{clean_FSC}
\alias{clean_FSC}
\title{Helper function}
\usage{
clean_FSC(FLS_vect, FSC, minsize = 4)
}
\arguments{
\item{FLS_vect}{A vector that contains the gene expression values of a sample.}

\item{FSC}{A nested list containing information about the regulons. For each
regulon it must contain MEMBERS, CONTRIBUTION, MOR, CONFIDENCE. A gene set
resource can be transformed into that format using the function df_to_list.}

\item{minsize}{A number that specifies the minimum number of genes that have
to be assigned to a regulon. Regulons with less genes are removed and not
included in the analysis.}
}
\value{
A nested list of regulons filtered according to minsize.
}
\description{
Helper function
}
