% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_gene_type}
\alias{convert_gene_type}
\title{Helper function}
\usage{
convert_gene_type(de_data, input_type, output_type)
}
\arguments{
\item{de_data}{Table with de results}

\item{input_type}{input gene IDs e.g. ENSEMBL}

\item{output_type}{output gene IDs e.g. Symbol}
}
\value{
Returns a df with gene names convert to symbols
}
\description{
Helper function to convert between commonly used gene IDs
(before committing I should create a resource separate from annotation dbi)
}
