% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR.R
\name{decouple}
\alias{decouple}
\title{decouple}
\usage{
decouple(mat, network, .source, .target, .options = list(), statistics)
}
\arguments{
\item{mat}{Evaluation matrix (e.g. expression matrix).
Target nodes in rows and samples in columns.}

\item{network}{Tibble or dataframe with edges and metadata.}

\item{.source}{Column in network with source nodes.}

\item{.target}{Column in network with target nodes.}

\item{.options}{Named list with edge attributes to use in the statistics.}

\item{statistics}{Statistical methods to be coupled.}
}
\value{
A long format tibble of the enrichment scores for each tf
 across the samples. Resulting tibble contains the following columns:
 \enumerate{
   \item{\code{tf}}: {Source nodes of \code{network}.}
   \item{\code{sample}}: {Samples representing each column of \code{mat}.}
   \item{\code{score}}: {Regulatory activity (enrichment score).}
   \item{\code{statistic}}: {Indicates which method is associated with which score.}
   \item{\code{metadata}}: {Metadata corresponding to the statistic collapsed to a string.}
 }
}
\description{
Calculate the TF activity per sample out of a gene expression matrix by
coupling a regulon network with a variety of statistics.
}
