% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR-package.R
\docType{package}
\name{decoupleR-package}
\alias{decoupleR}
\alias{decoupleR-package}
\title{decoupleR: decoupleR: Inferring biological activities from omics data using a collection of methods}
\description{
Computational methods allow the extraction of mechanistic signatures from omics data based on prior knowledge resources, reducing the dimensionality of the data for increased statistical power and better interpretability. Here, we present decoupleR, a Bioconductor package containing different statistical methods to extract these signatures within a unified framework. decoupleR allows the user to flexibly test any method with any resource. It incorporates methods that take into account the sign and weight of network interactions. Using decoupleR, we evaluated the performance of contemporary methods on transcriptomic and phospho-proteomic perturbation experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://saezlab.github.io/decoupleR/}
  \item Report bugs at \url{https://github.com/saezlab/decoupleR/issues}
}

}
\author{
\strong{Maintainer}: Pau Badia-i-Mompel \email{pau.badia@uni-heidelberg.de} (\href{https://orcid.org/0000-0002-1004-3923}{ORCID})

Authors:
\itemize{
  \item Jesús Vélez \email{jvelezmagic@gmail.com} (\href{https://orcid.org/0000-0001-5128-3838}{ORCID})
  \item Christian H. Holland \email{cholland2408@gmail.com} (\href{https://orcid.org/0000-0002-3060-5786}{ORCID})
}

}
\keyword{internal}
