% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eGSA.R
\name{df_to_list}
\alias{df_to_list}
\title{Create regulon list}
\usage{
df_to_list(genesets)
}
\arguments{
\item{genesets}{A data frame of gene sets including the following columns:
tf, confidence, target, mor and likelihood.}
}
\value{
A list of the regulons that can be used as FSC in SAFE_mean_scores().
}
\description{
This function converts the regulon data frame into a list needed for the mean
statistics.
}
