% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signscore.R
\name{directed2singscore}
\alias{directed2singscore}
\title{Helper function}
\usage{
directed2singscore(genesets, minsize)
}
\arguments{
\item{genesets}{Directed genesets (e.g. dorothea, progeny)}

\item{minsize}{Minimum number of genes/targets in the geneset}
}
\value{
A list of two lists: one for positively and one list for negatively
regulated genesets with gene/target character vectors suitable for signscore
}
\description{
Helper function to convert a Directed/signed gene set to two seperate
gene sets according to the direction in which they regulate the genes
suitable for the \code{\link[=singscore]{singscore::simpleScore()}} function.
}
\keyword{internal}
