% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{dorothea2gsea}
\alias{dorothea2gsea}
\title{Helper function}
\usage{
dorothea2gsea(genesets)
}
\arguments{
\item{genesets}{dorothea gene sets}
}
\value{
dorothea gene sets suitable for gsea statistic
}
\description{
Helper function to convert DoRothEA gene sets to standardized gene sets
suitable for the \code{\link[=fgsea]{fgsea::fgsea()}} function.
}
