% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-viper.R
\name{dorothea2viper}
\alias{dorothea2viper}
\title{Helper function}
\usage{
dorothea2viper(genesets)
}
\arguments{
\item{genesets}{dorothea gene sets}
}
\value{
dorothea gene sets suitable for viper statistic
}
\description{
Helper function to convert DoRothEA gene sets to standardized gene sets
suitable for the \code{\link[=viper]{viper::viper()}} function.
}
\keyword{internal}
