% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-decoupler-formats.R
\name{.decoupler_mat_format}
\alias{.decoupler_mat_format}
\title{DecoupleR mat format}
\arguments{
\item{mat}{Evaluation matrix (e.g. expression matrix).
Target nodes in rows and conditions in columns.}
}
\description{
DecoupleR mat format
}
\keyword{internal}
