% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-decoupler-formats.R
\name{.decoupler_network_format}
\alias{.decoupler_network_format}
\title{DecoupleR network format}
\arguments{
\item{network}{Tibble or dataframe with edges and it's associated metadata.}

\item{.source}{Column with source nodes.}

\item{.target}{Column with target nodes.}

\item{.mor}{Column with edge mode of regulation (i.e. mor).}

\item{.likelihood}{Column with edge likelihood.}
}
\description{
A network passed to any \code{run_} method in the package must contain at
least two attributes: \code{.source} and \code{.target}. In addition,
the methods must map their corresponding metadata associated with their edges.
}
\keyword{internal}
