% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataset-converters.R
\name{extract_sets}
\alias{extract_sets}
\title{Extract sets}
\usage{
extract_sets(network)
}
\arguments{
\item{network}{Tibble or dataframe with edges and it's associated metadata.}
}
\description{
Extracts feature sets from a renamed network (see \link{rename_net}).
}
\examples{
inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")
mat <- readRDS(file.path(inputs_dir, "mat.rds"))
net <- readRDS(file.path(inputs_dir, "net.rds"))
net <- rename_net(net, source, target, mor)
extract_sets(net)
}
