% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR-pre.R
\name{filt_minsize}
\alias{filt_minsize}
\title{Filter sources with minsize targets}
\usage{
filt_minsize(mat_f_names, network, minsize = 5)
}
\arguments{
\item{mat_f_names}{Feature names of mat.}

\item{network}{Tibble or dataframe with edges and it's associated metadata.}
}
\value{
Filtered network.
}
\description{
Filter sources of a net with less than minsize targets
}
