% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bench-run.R
\name{filter_sets}
\alias{filter_sets}
\title{Helper Function to filter and format the gene set resource}
\usage{
filter_sets(set_source, gene_source, .lvls, lvls, .minsize)
}
\description{
Helper Function to filter and format the gene set resource
}
