% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bench-run.R
\name{format_design}
\alias{format_design}
\title{Helper Function to to generate the booleans used to check if the current
locations/data objects are the same as the previous one}
\usage{
format_design(design)
}
\arguments{
\item{design}{location of a json file with run design specifications}
}
\value{
a design tibble to be used in benchmarking
}
\description{
Helper Function to to generate the booleans used to check if the current
locations/data objects are the same as the previous one
}
