% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench-summ.R
\name{format_roc}
\alias{format_roc}
\title{Helper function to format (Precision-Recall) Receiver Operator Curve results}
\usage{
format_roc(.res_tibble, roc_column)
}
\arguments{
\item{.res_tibble}{formatted bench result tibble with added AUROC column}

\item{roc_column}{PR/ROC column to format}
}
\value{
returns
}
\description{
Helper function to format (Precision-Recall) Receiver Operator Curve results
}
