% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signscore.R
\name{generateNull_bidirect}
\alias{generateNull_bidirect}
\title{Function to generate null hypothesis against which the single sample scores
of signscore will be compared (Directed genesets with + and - signs)}
\usage{
generateNull_bidirect(genesets_up, genesets_dn, rankData, perm_num, ncores)
}
\arguments{
\item{genesets_up}{A character vector of positively reg. gene set IDs}

\item{genesets_dn}{A character vector of negatively reg. gene set IDs}

\item{rankData}{Ranked expression matrix}

\item{perm_num}{number of permutations to be performed}

\item{ncores}{number of cores to be used for the calculation}
}
\value{
permutation results to be used to perform a one-tailed test
}
\description{
Refer to the \code{\link[=singscore]{singscore::generateNull()}} function.
}
\keyword{internal}
