% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signscore.R
\name{generateNull_undirected}
\alias{generateNull_undirected}
\title{Function to generate null hypothesis against which the single sample scores
of signscore will be compared (Undirected genesets with + OR - signs)}
\usage{
generateNull_undirected(geneset_members, rankData, perm_num, ncores)
}
\arguments{
\item{geneset_members}{A character vector of gene IDs of the gene set}

\item{rankData}{Ranked expression matrix}

\item{perm_num}{number of permutations to be performed}

\item{ncores}{number of cores to be used for the calculation}
}
\value{
permutation results to be used to perform a one-tailed test
}
\description{
Refer to the \code{\link[=singscore]{singscore::generateNull()}} function.
}
\keyword{internal}
