% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench-summ.R
\name{get_auroc_heat}
\alias{get_auroc_heat}
\title{Helper function to produce AUROC heatmap}
\usage{
get_auroc_heat(auroc_tibble)
}
\arguments{
\item{auroc_tibble}{Tibble with calculated AUROC}
}
\value{
returns an AUROC or Precision-Recall AUC heatmap
}
\description{
Helper function to produce AUROC heatmap
}
