% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench-summ.R
\name{get_bench_summary}
\alias{get_bench_summary}
\title{Function that provides summary and plots for the benchmark run}
\usage{
get_bench_summary(.res_tibble)
}
\arguments{
\item{.res_tibble}{formatted bench result tibble with added auroc column}
}
\value{
AUROC summary with TF coverage, ROC, AUROC, PRROC, Run time,
ROC plots, and Heatmap plots
}
\description{
Function that provides summary and plots for the benchmark run
}
