% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-omnipath.R
\name{get_collectri}
\alias{get_collectri}
\title{CollecTRI gene regulatory network.
Wrapper to access CollecTRI gene regulatory network. CollecTRI is a
comprehensive resource containing a curated collection of transcription
factors (TFs) and their target genes. It is an expansion of DoRothEA.
Each interaction is weighted by its mode of regulation (either positive or negative).}
\usage{
get_collectri(organism = "human", split_complexes = FALSE, ...)
}
\arguments{
\item{organism}{Which organism to use. Only human, mouse and rat are available.}

\item{split_complexes}{Whether to split complexes into subunits. By default
complexes are kept as they are.}

\item{...}{Ignored.}
}
\description{
CollecTRI gene regulatory network.
Wrapper to access CollecTRI gene regulatory network. CollecTRI is a
comprehensive resource containing a curated collection of transcription
factors (TFs) and their target genes. It is an expansion of DoRothEA.
Each interaction is weighted by its mode of regulation (either positive or negative).
}
\examples{
collectri <- get_collectri(organism='human', split_complexes=FALSE)
}
