% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{get_config_filename_from_dir}
\alias{get_config_filename_from_dir}
\title{Return config file path found in a directory}
\usage{
get_config_filename_from_dir(path)
}
\arguments{
\item{path}{The directory were config files are looking for.}
}
\value{
the full path of the config file found.
If no config file are found, it returns a `NULL`.
}
\description{
Return config file path found in a directory
}
