% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{get_CONTmatrix}
\alias{get_CONTmatrix}
\title{Helper function}
\usage{
get_CONTmatrix(FSC, target_features)
}
\arguments{
\item{FSC}{A nested list containing information about the regulons. For each
regulon it must contain MEMBERS, CONTRIBUTION, MOR, CONFIDENCE. A gene set
resource can be transformed into that format using the function df_to_list.}

\item{target_features}{A vector including the target genes of a tf.}
}
\value{
A matrix including the contribution of a tf on its targets
}
\description{
Helper function
}
