% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-omnipath.R
\name{get_KSN_omnipath}
\alias{get_KSN_omnipath}
\title{Omnipath KSN}
\usage{
get_KSN_omnipath(...)
}
\arguments{
\item{...}{Passed to \code{OmnipathR::import_omnipath_enzsub}.}
}
\description{
Generate a ready to use, curated Kinase/Substrate Network from omnipath
}
\details{
Import PTM network from omnipath, then filter out anything that is not
phospho or dephosphorilation. Then format the columns for use with decoupleR
functions.
}
