% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-omnipath.R
\name{get_progeny}
\alias{get_progeny}
\title{Pathway RespOnsive GENes for activity inference (PROGENy).}
\usage{
get_progeny(organism = "human", top = 500)
}
\arguments{
\item{organism}{Which organism to use. Only human and mouse are available.}

\item{top}{Number of genes per pathway to return.}
}
\description{
Wrapper to access PROGENy model gene weights. Each pathway is defined with a
collection of target genes, each interaction has an associated p-value and
weight. The top significant interactions per pathway are returned.
}
\examples{
progeny <- get_progeny(organism='human', top=500)
}
