% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR-pre.R
\name{get_toy_data}
\alias{get_toy_data}
\title{Generate a toy \code{mat} and \code{network}.}
\usage{
get_toy_data(n_samples = 24, seed = 42)
}
\arguments{
\item{n_samples}{Number of samples to simulate.}

\item{seed}{A single value, interpreted as an integer, or NULL for random
number generation.}
}
\value{
List containing \code{mat} and \code{network}.
}
\description{
Generate a toy \code{mat} and \code{network}.
}
\examples{
data <- get_toy_data()
mat <- data$mat
network <- data$network
}
