% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR-pre.R
\name{intersect_regulons}
\alias{intersect_regulons}
\title{Intersect network target genes with expression matrix.}
\usage{
intersect_regulons(mat, network, source, target, minsize)
}
\arguments{
\item{network}{Tibble or dataframe with edges and it's associated metadata.}

\item{source}{Column with source nodes.}

\item{target}{Column with target nodes.}

\item{.target}{Maximum number of targets allowed per regulon.}
}
\value{
Filtered tibble.
}
\description{
Keep only edges which its target genes belong to the expression matrix.
}
\examples{
inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")
mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))
filter_regulons(mat, network, target)
}
