% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_undefined}
\alias{is_undefined}
\title{Test if object if of class _UNDEFINED}
\usage{
is_undefined(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a boolean
}
\description{
Test if object if of class _UNDEFINED
}
