% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{make_gsea_genesets}
\alias{make_gsea_genesets}
\title{Make gene sets for GSEA}
\usage{
make_gsea_genesets(genesets)
}
\arguments{
\item{genesets}{A dataframe of gene set that must contain the
columns \code{geneset} and \code{gene}.}
}
\value{
Gene sets in the \code{\link[=fgsea]{fgsea::fgsea()}} format.
}
\description{
This function convert gene sets in a table format to the format required by
the \code{\link[=viper]{fgsea::fgsea}} function.
}
