% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper.R
\name{make_viper_genesets}
\alias{make_viper_genesets}
\title{Make gene sets for VIPER}
\usage{
make_viper_genesets(genesets)
}
\arguments{
\item{genesets}{A dataframe of gene sets that must contain the
columns \code{geneset}, \code{gene}, \code{mor} and \code{likelihood}.}
}
\value{
Gene sets in the \code{\link[=viper]{viper::viper()}} format.
}
\description{
This function convert gene sets in a table format to the format required by
the \code{\link[=viper]{viper::viper()}} function.
}
\keyword{internal}
