% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{progeny2gsea}
\alias{progeny2gsea}
\title{Helper function}
\usage{
progeny2gsea(genesets)
}
\arguments{
\item{genesets}{progeny gene sets}
}
\value{
progeny gene sets suitable for gsea statistic
}
\description{
Helper function to convert PROGENy gene sets to standardized gene sets
suitable for the \code{\link[=fgsea]{fgsea::fgsea()}} function.
}
