% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper.R
\name{progeny2viper}
\alias{progeny2viper}
\title{Helper function}
\usage{
progeny2viper(genesets)
}
\arguments{
\item{genesets}{progeny gene sets}
}
\value{
progeny gene sets suitable for viper statistic
}
\description{
Helper function to convert PROGENy gene sets to standardized gene sets
suitable for the \code{\link[=viper]{viper::viper()}} function.
}
\keyword{internal}
