% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_string}
\alias{random_string}
\title{Generate a random string of specified length}
\usage{
random_string(string_length = 6, replace = TRUE)
}
\arguments{
\item{string_length}{Integer, length of the string to generate.}

\item{replace}{Bollean, Use replace in the sampling procedure.}
}
\value{
A random string.
}
\description{
Generate a random string of specified length
}
