% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_matrix.R
\name{randomize_matrix}
\alias{randomize_matrix}
\title{Randomize matrix}
\usage{
randomize_matrix(mat, randomize_type = c("rows", "cols_independently"))
}
\arguments{
\item{mat}{Matrix to randomize.}

\item{randomize_type}{How to randomize.}
}
\value{
Randomized matrix
}
\description{
Randomize matrix
}
