% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signscore.R
\name{regnetwork2viper}
\alias{regnetwork2viper}
\title{Helper function}
\usage{
regnetwork2viper(genesets)
}
\arguments{
\item{genesets}{RegNetwork gene sets}
}
\value{
RegNetwork gene sets suitable for singscore statistic
}
\description{
Helper function to convert RegNetwork gene sets to standardized gene sets
\code{\link[=singscore]{singscore::simpleScore()}} function.
}
\keyword{internal}
