% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_comments}
\alias{remove_comments}
\title{Parse config file and remove comment lines}
\usage{
remove_comments(lines)
}
\arguments{
\item{lines}{A list of config file lines.}
}
\value{
A list of config file lines where comment lines have been filtered
out.
}
\description{
Parse config file and remove comment lines
}
