% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_empty_lines}
\alias{remove_empty_lines}
\title{Parse config file and remove empty lines}
\usage{
remove_empty_lines(lines)
}
\arguments{
\item{lines}{A list of config file lines.}
}
\value{
A list of config file lines where empty lines have been filtered out.
}
\description{
Parse config file and remove empty lines
}
