% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataset-converters.R
\name{rename_defaults}
\alias{rename_defaults}
\title{Rename and add column with defaults}
\usage{
rename_defaults(.data, ..., .def_col_val = c())
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{For \code{rename()}: <\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Use
\code{new_name = old_name} to rename selected variables.

For \code{rename_with()}: additional arguments passed onto \code{.fn}.}

\item{.def_col_val}{Named vector with columns with default values
if none exist after rename.}
}
\value{
An object of the same type as \code{.data}. The output has the following
properties:
\itemize{
\item Rows are not affected.
\item Column names are changed; column order is preserved.
\item Data frame attributes are preserved.
\item Groups are updated to reflect new names.
}
}
\description{
\code{rename_defaults} combine the \code{\link[dplyr]{rename}} way of
working and with the \code{\link[tibble]{add_column}} to add columns
with default values in case they don't exist after renaming the dataset.
}
\keyword{internal}
