% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-aucell.R
\name{run_aucell}
\alias{run_aucell}
\title{AUCell}
\usage{
run_aucell(
  mat,
  network,
  .source = .data$source,
  .target = .data$target,
  aucMaxRank = ceiling(0.05 * nrow(rankings)),
  nproc = 4,
  seed = 42
)
}
\arguments{
\item{mat}{Matrix to evaluate (e.g. expression matrix).
Target nodes in rows and conditions in columns.
\code{rownames(mat)} must have at least one intersection with the elements
in \code{network} \code{.target} column.}

\item{network}{Tibble or dataframe with edges and it's associated metadata.}

\item{.source}{Column with source nodes.}

\item{.target}{Column with target nodes.}

\item{aucMaxRank}{Threshold to calculate the AUC.}

\item{nproc}{Number of cores to use for computation.}
}
\description{
Calculates regulatory activities using Area Under the Curve (AUC) from AUCell
}
\details{
This function is a wrapper for the method \code{AUCell}. It uses the
"Area Under the Curve" (AUC) to calculate whether a critical subset of input
molecular features is enriched for each sample.
}
\examples{
inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

run_aucell(mat, network, .source='tf')
}
\seealso{
Other decoupleR statistics: 
\code{\link{decouple}()},
\code{\link{run_fgsea}()},
\code{\link{run_gsva}()},
\code{\link{run_mdt}()},
\code{\link{run_mlm}()},
\code{\link{run_ora}()},
\code{\link{run_udt}()},
\code{\link{run_ulm}()},
\code{\link{run_viper}()},
\code{\link{run_wmean}()},
\code{\link{run_wsum}()}
}
\concept{decoupleR statistics}
