% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR-consensus.R
\name{run_consensus}
\alias{run_consensus}
\title{Function to generate a consensus score between methods from the
result of decouple}
\usage{
run_consensus(
  df,
  condition = "condition",
  statistic = "statistic",
  include_time = FALSE
)
}
\arguments{
\item{df}{decouple results}

\item{condition}{Column name for sample names}

\item{statistic}{Column name for statistic names}
}
\value{
Updated tibble with the computed consensus score bewteen methods
}
\description{
Function to generate a consensus score between methods from the
result of decouple
}
