% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eGSA.R
\name{run_eGSA}
\alias{run_eGSA}
\title{eGSA}
\usage{
run_eGSA(emat, genesets, statistic, options = list())
}
\arguments{
\item{emat}{An expression matrix with genes (HGNC symbol) in rows and samples
in columns.}

\item{genesets}{A data frame of gene sets. The structure is dependent on the
gene set resource.}

\item{statistic}{A character indicating which of the following statistics
should be used to calculate the activity: mean, normalized_mean, GSEA}

\item{options}{A list of named options to pass to
SAFE_mean() such as \code{minsize} or \code{mixed}.}
}
\value{
A matrix of normalized enrichment scores for each gene set across all
 samples.
}
\description{
This function is a convenient wrapper for using one of the eGSA statistics
in decoupleR.
}
