% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{run_gsea}
\alias{run_gsea}
\title{GSEA wrapper}
\usage{
run_gsea(emat, genesets, options = list(nperm = 1000), gs_resource, tidy = F)
}
\arguments{
\item{emat}{An expression matrix with genes (HGNC symbol) in rows and samples
in columns.}

\item{genesets}{A data frame of gene sets that must contain at least the
columns \code{geneset} and \code{gene}}

\item{options}{A list of named options to pass to
\code{\link[=fgsea]{fgsea::fgsea()}} such as \code{nperm} or \code{minSize},
while \code{nperm} must be specified. These options should not include
\code{pathways} or \code{stats}.}

\item{gs_resource}{A character indicating the gene set resource. Based on
this argument \code{genesets} is processed futher via diverse helper
functions (e.g. \code{\link[=dorothea2gsea]{dorothea2gsea()}})}

\item{tidy}{Logical, whether computed normalized enrichment scores should be
returned in a tidy format.}
}
\value{
A matrix of normalized enrichment score for each gene sets across all
 samples. If \code{tidy} is TRUE the results are retured in a tidy format,
 including also p-values and leading edge genes.
}
\description{
This function is a convenient wrapper for the
\code{\link[=fgsea]{fgsea::fgsea()}} function to perform GSEA
}
