% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-mlm.R
\name{run_mlm}
\alias{run_mlm}
\title{Multivariate Linear Model (MLM)}
\usage{
run_mlm(
  mat,
  network,
  .source = .data$source,
  .target = .data$target,
  .mor = .data$mor,
  .likelihood = .data$likelihood,
  sparse = FALSE,
  center = FALSE,
  na.rm = FALSE,
  minsize = 5,
  regularization = NULL
)
}
\arguments{
\item{mat}{Matrix to evaluate (e.g. expression matrix).
Target nodes in rows and conditions in columns.
\code{rownames(mat)} must have at least one intersection with the elements
in \code{network} \code{.target} column.}

\item{network}{Tibble or dataframe with edges and it's associated metadata.}

\item{.source}{Column with source nodes.}

\item{.target}{Column with target nodes.}

\item{.mor}{Column with edge mode of regulation (i.e. mor).}

\item{.likelihood}{Deprecated argument. Now it will always be set to 1.}

\item{sparse}{Deprecated parameter.}

\item{center}{Logical value indicating if \code{mat} must be centered by
\code{\link[base:colSums]{base::rowMeans()}}.}

\item{na.rm}{Should missing values (including NaN) be omitted from the
calculations of \code{\link[base:colSums]{base::rowMeans()}}?}

\item{minsize}{Integer indicating the minimum number of targets per source.}

\item{regularization}{Indicate whether apply regularization. NULL if no regularization. Set to 'ridge' or 'lasso' to use the corresponding regularization scheme implemented in glmnet. Can also set to numerical values between 0 (ridge) to 1 (lasso), and any value in between represent elastic net.}
}
\value{
A long format tibble of the enrichment scores for each source
across the samples. Resulting tibble contains the following columns:
\enumerate{
\item \code{statistic}: Indicates which method is associated with which score.
\item \code{source}: Source nodes of \code{network}.
\item \code{condition}: Condition representing each column of \code{mat}.
\item \code{score}: Regulatory activity (enrichment score).
}
}
\description{
Calculates regulatory activities by fitting multivariate linear models (MLM)
}
\details{
MLM fits a multivariate linear model to estimate regulatory activities.
MLM transforms a given network into an adjacency matrix, placing sources as
columns and targets as rows. The matrix is filled with the associated weights
for each interaction. This matrix is used to fit a linear model to predict
the observed molecular readouts per sample. The obtained t-values from the
fitted model are the activities of the regulators.

MLM also provides the option to fit a regularized multivariate linear model
(ridge or lasso). Using the functions in the "glmnet" package, different
stringency of regularization (lambda) results in an ensemble of many solutions.
At each solution, the coefficients of regulators (sources) are scaled by the
absolute sum of all coefficients, representing the "importance" of each regulator.
For each regulator, its optimal "importance" (strongly positive or negative)
among the ensemble is used to represent its activity.
}
\examples{
inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

run_mlm(mat, network, .source='tf')
}
\seealso{
Other decoupleR statistics: 
\code{\link{decouple}()},
\code{\link{run_aucell}()},
\code{\link{run_fgsea}()},
\code{\link{run_gsva}()},
\code{\link{run_mdt}()},
\code{\link{run_ora}()},
\code{\link{run_udt}()},
\code{\link{run_ulm}()},
\code{\link{run_viper}()},
\code{\link{run_wmean}()},
\code{\link{run_wsum}()}
}
\concept{decoupleR statistics}
