% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscira.R
\name{run_pscira}
\alias{run_pscira}
\title{PSCIRA (Permutation Single Cell Inference of Regulatory Activity)}
\usage{
run_pscira(
  emat,
  regulons,
  .source = .data$tf,
  .target = .data$target,
  .target_profile = .data$mor,
  .sparse = TRUE,
  times = 10,
  seed = 42
)
}
\arguments{
\item{emat}{A named expression matrix. E.g Genes in rows and samples in columns.}

\item{regulons}{A data frame of regulons in table format.}

\item{.source}{Column name in regulons with the factors to calculate the enrichment scores.}

\item{.target}{Column name in regulons that relates the labels of the rows in the expression matrix.}

\item{.target_profile}{Column name in regulons that contains the target profile. E.g. MoR.}

\item{.sparse}{Logical value indicating if the generated profile matrix should be sparse.}

\item{times}{Number of replications.}

\item{seed}{A single value, interpreted as an integer, or NULL.}
}
\value{
A long format tibble of the enrichment results for each set of genes
 across the samples. Resulting tibble contains the following columns:
 \enumerate{
   \item \code{source} Resources from the \code{.source} column of the \code{regulons} data frame.
   \item \code{condition} Samples representing each column of \code{emat}.
   \item \code{score} Regulatory activity of each resource for each condition.
 }
}
\description{
Calculate the regulatory activity of each tf by multiplying the expression
values of its objectives with their corresponding associated profiles for
each given condition.The result is equal to the z-score of the found value
compared to its null distribution.
}
\details{
Estimation of regulatory activity: A linear regression of the expression profile
is performed against the "target profile" of the given TF, where in the target
profile, any member of the regulon will be assigned a +1 for activated actions,
a -1 for inhibitory activations and, finally, a 0 for all genes that are not
a member of the TF regulon. TF activity is then defined as the t-statistic of
this linear regression.
}
