% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signscore.R
\name{run_singscore}
\alias{run_singscore}
\title{SingScore Wrapper}
\usage{
run_singscore(
  expr_matrix,
  tiesMethod,
  genesets,
  resource_name,
  minsize,
  perm_num,
  ncores,
  directed,
  permutate
)
}
\arguments{
\item{expr_matrix}{Gene expression matrix}

\item{tiesMethod}{Methods to use when ranking ties: min, max, average}

\item{genesets}{Gene set resource}

\item{resource_name}{A character indicating the gene set resource. Based on
this argument \code{genesets} is processed futher via diverse helper
functions (e.g. \code{\link[=dorothea2viper]{dorothea2viper()}} and
 \code{\link[=directed2singscore]{directed2singscore()}}).}

\item{minsize}{Minimum number of genes/targets in the gene set}

\item{perm_num}{Number of permutations to be performed}

\item{ncores}{Number of cores to be used for the calculation}

\item{directed}{Bool: Use to indicated whether the Dataset Resource counts
directed genesets/pathways or not}

\item{permutate}{Bool: Used to indicate whether we wish to obtain p-values or
just singscores per sample}
}
\value{
A dataframe with p-values for each sample or a list of singscore
dataframes (if permute is TRUE) or a list of dataframes with signscore and
dispersion (if permute is FALSE)
}
\description{
SingScore Wrapper
}
\details{
Please note that a MultiScore method is exists within singscore
which could be used to produce a DF with scores from a GeneSetCollection,
but this is not feasible because generateNull still has to be applied to
each geneset. Also, all genesets must be directed for the GeneSetCollection
to work with all of them or it would again require a split in un/directed gs
}
\examples{
# Run singscore with dorothea
dorothea_singscore <- run_singscore(expr_matrix,
"min",
dorothea,
"dorothea",
4,
100,
6,
TRUE,
TRUE)

# Run singscore with Regnetwork
regnetwork_singscore <- run_singscore(expr_matrix,
                                      "min",
                                      regnetwork,
                                      "regnetwork",
                                      4,
                                      100,
                                      6,
                                      FALSE,
                                      TRUE)

# Run singscore with PROGENy
progeny_singscore <- run_singscore(expr_matrix,
                                  "min",
                                   progeny,
                                   "progeny",
                                   4,
                                   100,
                                   6,
                                   TRUE,
                                   TRUE)


}
