% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-wmean.R
\name{run_wmean}
\alias{run_wmean}
\title{Weighted mean}
\usage{
run_wmean(
  mat,
  network,
  .source = .data$source,
  .target = .data$target,
  .mor = .data$mor,
  .likelihood = .data$likelihood,
  times = 100,
  seed = 42,
  sparse = TRUE,
  randomize_type = "rows"
)
}
\arguments{
\item{mat}{Matrix to evaluate (e.g. expression matrix).
Target nodes in rows and conditions in columns.
\code{rownames(mat)} must have at least one intersection with the elements
in \code{network} \code{.target} column.}

\item{network}{Tibble or dataframe with edges and it's associated metadata.}

\item{.source}{Column with source nodes.}

\item{.target}{Column with target nodes.}

\item{.mor}{Column with edge mode of regulation (i.e. mor).}

\item{.likelihood}{Column with edge likelihood.}

\item{times}{How many permutations to do?}

\item{seed}{A single value, interpreted as an integer, or NULL for random
number generation.}

\item{sparse}{Should the matrices used for the calculation be sparse?}

\item{randomize_type}{How to randomize the expression matrix.}
}
\value{
A long format tibble of the enrichment scores for each source
across the samples. Resulting tibble contains the following columns:
\enumerate{
\item \code{statistic}: Indicates which method is associated with which score.
\item \code{source}: Source nodes of \code{network}.
\item \code{condition}: Condition representing each column of \code{mat}.
\item \code{score}: Regulatory activity (enrichment score).
\item \code{p_value}: p-value for the score of the method.
}
}
\description{
Calculate the activity of all regulons in \code{network} through the conditions in
the \code{mat} matrix by calculating the weighted mean over the expression of all genes.
}
\details{
\code{run_wmean()} calculates the activity score, but in addition, it takes
advantage of the permutations used to calculate the \code{p-value}, to provide
the normalized and corrected activity scores. This is represented in the \code{statistic} column
which will contain three values for each call to \code{run_wmean()}; \strong{wmean},
\strong{norm_wmean} and \strong{corr_wmean}.
}
\examples{
inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

run_wmean(mat, network, .source='tf')
}
\seealso{
Other decoupleR statistics: 
\code{\link{decouple}()},
\code{\link{run_aucell}()},
\code{\link{run_gsva}()},
\code{\link{run_nolea}()},
\code{\link{run_ora}()},
\code{\link{run_scira}()},
\code{\link{run_viper}()},
\code{\link{run_wsum}()}
}
\concept{decoupleR statistics}
