% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{SAFE_mean_normalized}
\alias{SAFE_mean_normalized}
\title{SAFE MEAN normalized}
\usage{
SAFE_mean_normalized(
  FLS_vect,
  FSC,
  mixed = FALSE,
  null_model = NULL,
  nperm = 100,
  minsize = 4
)
}
\arguments{
\item{FLS_vect}{A vector that contains the gene expression values of a sample.}

\item{FSC}{A nested list containing information about the regulons. For each
regulon it must contain MEMBERS, CONTRIBUTION, MOR, CONFIDENCE. A gene set
resource can be transformed into that format using the function df_to_list.}

\item{mixed}{Logical, if TRUE permutations are performed only within the
positive and negative values. Default is FALSE.}

\item{null_model}{Define a null model. If NULL (default) the null model is
generated by the permutations.}

\item{nperm}{A number that defines how many permutations are performed.}

\item{minsize}{A number that specifies the minimum number of genes that have
to be assigned to a regulon. Regulons with less genes are removed and not
included in the analysis.}
}
\value{
A matrix of normalized enrichment scores for each gene set across all
 samples.
}
\description{
This function calculates the regulon activity of one sample by calculating
a normalized mean using the z-scores.
}
