% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR-pre.R
\name{show_methods}
\alias{show_methods}
\title{Show methods}
\usage{
show_methods()
}
\description{
Prints the methods available in decoupleR. The first column correspond to
the function name in decoupleR and the second to the method's full name.
}
\examples{
show_methods()
}
