% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-omnipath.R
\name{show_resources}
\alias{show_resources}
\title{Shows available resources in Omnipath. For more information visit the
official website for \href{https://omnipathdb.org/}{Omnipath}.}
\usage{
show_resources()
}
\description{
Shows available resources in Omnipath. For more information visit the
official website for \href{https://omnipathdb.org/}{Omnipath}.
}
\examples{
decoupleR::show_resources()
}
