% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signscore.R
\name{simpleScore_bidir}
\alias{simpleScore_bidir}
\title{Function to calculate singScore and pvalues for bidirectional sets
which implements the \code{\link[=singscore]{singscore::simpleScore()}} and
implements the \code{\link[=singscore]{singscore::getPvals()}} functions.}
\usage{
simpleScore_bidir(gs_resource, geneset, rankData, perm_num, ncores)
}
\arguments{
\item{gs_resource}{gene sets with a list of
named lists with positively and negatively regulated genes}

\item{geneset}{GO/TF/Pathway gene set}

\item{rankData}{ranked expression matrix}

\item{perm_num}{number of permutations to be performed}

\item{ncores}{number of cores to be used}
}
\value{
A matrix of p-values
}
\description{
Function to calculate singScore and pvalues for bidirectional sets
which implements the \code{\link[=singscore]{singscore::simpleScore()}} and
implements the \code{\link[=singscore]{singscore::getPvals()}} functions.
}
\keyword{internal}
