% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_logical.R
\name{to_logical}
\alias{to_logical}
\title{Convert categorical representations of true/false values to a logical}
\usage{
to_logical(x)
}
\arguments{
\item{x}{a vector of boolean compatible values.}
}
\value{
a vector of boolean
}
\description{
Allow the convertion of logical related data to actual Boolean
}
\examples{
to_logical(c("yes", "no"))
}
