% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench-format.R
\name{bench_format}
\alias{bench_format}
\title{Function to format benchmarking results}
\usage{
bench_format(bench_res, .silent)
}
\arguments{
\item{bench_res}{benchmarking results}

\item{.silent}{bool whether to silence warnings or not}
}
\value{
formatted benchmarking results
}
\description{
Function to format benchmarking results
}
\details{
If infinite values are present in the results, this function will
notify the user.
}
