% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-bench_res.R
\docType{class}
\name{BenchResult-class}
\alias{BenchResult-class}
\title{S4 Class used to format benchmark output.}
\description{
S4 Class used to format benchmark output.
}
\section{Fields}{

\describe{
\item{\code{bench_res}}{Formatted or non-formatted Benchmark output}

\item{\code{summary}}{Summary returned by the bench_sumplot functions - it contains
a summary table, roc plot, precision-recall curve (prc) plot, auroc heatmap,
and precision-recall area under the curve heatmap}

\item{\code{design}}{The input design tibble used to generate the benchmark results}
}}

