% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench-auroc.R
\name{calc_curve}
\alias{calc_curve}
\title{This function takes the elements of the \code{activity} column and calculates
precision-recall and ROC curves (depending on \code{curve}).
The \code{activity} column is populated with the output for each stat method and
results from the \code{run_benchmark()} function. Each of the elements
in \code{activity} are results from runs of the \link{decouple} wrapper.}
\usage{
calc_curve(
  df,
  downsampling = FALSE,
  times = 1000,
  ranked = FALSE,
  curve = "ROC"
)
}
\arguments{
\item{df}{run_benchmark roc column provided as input}

\item{downsampling}{logical flag indicating if the number of TN should be
downsampled to the number of TP}

\item{times}{integer showing the number of downsampling}

\item{ranked}{logical flag indicating if input is derived from composite
ranking that already took up-/downregulation (sign) into account}

\item{curve}{whether to return a Precision-Recall Curve ("PR") or ROC ("ROC")}
}
\value{
tidy data frame with precision, recall, auc, n, tp, tn and coverage
in the case of PR curve; or sensitivity and specificity, auc, n, tp, tn and
coverage in the case of ROC
}
\description{
This function takes the elements of the \code{activity} column and calculates
precision-recall and ROC curves (depending on \code{curve}).
The \code{activity} column is populated with the output for each stat method and
results from the \code{run_benchmark()} function. Each of the elements
in \code{activity} are results from runs of the \link{decouple} wrapper.
}
