% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bench-auroc.R
\name{calc_pr_curve}
\alias{calc_pr_curve}
\title{This function calculates precision recall curves}
\usage{
calc_pr_curve(df)
}
\arguments{
\item{df}{decouple output (activity element from benchmark result)}
}
\value{
tidy df containing recall, precision, auc, tp, tn and coverage
}
\description{
This function calculates precision recall curves
}
