% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bench-run.R
\name{check_prereq}
\alias{check_prereq}
\title{Helper Function that checks if the  preceding vector element is the same
as the current element}
\usage{
check_prereq(vector_loc)
}
\arguments{
\item{vector_loc}{char vector with directories}
}
\value{
logical values describing whether the location of the loaded files
has changes
}
\description{
Helper Function that checks if the  preceding vector element is the same
as the current element
}
