% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataset-converters.R
\name{convert_to_}
\alias{convert_to_}
\alias{convert_to_scira}
\alias{convert_to_pscira}
\alias{convert_to_mean}
\alias{convert_to_viper}
\alias{convert_to_gsva}
\title{Convert a network to run under the method of interest.}
\usage{
convert_to_(network)

convert_to_scira(network, .source, .target, .mor = NULL)

convert_to_pscira(network, .source, .target, .mor = NULL)

convert_to_mean(network, .source, .target, .mor = NULL, .likelihood = NULL)

convert_to_viper(network, .source, .target, .mor = NULL, .likelihood = NULL)

convert_to_gsva(network, .source, .target)
}
\arguments{
\item{network}{Tibble or dataframe with edges and it's associated metadata.}

\item{.source}{Column with source nodes.}

\item{.target}{Column with target nodes.}

\item{.mor}{Column with edge mode of regulation (i.e. mor).}

\item{.likelihood}{Column with edge likelihood.}
}
\value{
\itemize{
\item \code{convert_to_}
Return same as input.
}
\itemize{
\item \code{convert_to_gsva()}
Return a list of regulons suitable for \code{\link[GSVA:gsva]{GSVA::gsva()}}.
\item \code{convert_to_mean()}
Return a tibble with four columns: \code{tf}, \code{target}, \code{mor} and \code{likelihood}.
\item \code{convert_to_pscira()}
Returns a tibble with three columns: \code{tf}, \code{target} and \code{mor}.
\item \code{convert_to_scira()}
Returns a tibble with three columns: \code{tf}, \code{target} and \code{mor}.
\item \code{convert_to_viper()}
Return a list of regulons suitable for \code{\link[viper:viper]{viper::viper()}}
}
}
\description{
Convert a long-format network to the suggested standard for the
specified \verb{run_\{statistic\}()}. If the default parameters are not modified,
then the function sets its own null values for those columns.
}
\examples{
inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

convert_to_(network)
convert_to_gsva(network, tf, target)
convert_to_mean(network, tf, target, mor, likelihood)
convert_to_pscira(network, tf, target, mor)
convert_to_scira(network, tf, target, mor)
convert_to_viper(network, tf, target, mor, likelihood)
}
\seealso{
\code{\link[=convert_f_defaults]{convert_f_defaults()}}
}
\concept{convert_to_ variants}
