% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataset-converters.R
\name{convert_to_}
\alias{convert_to_}
\alias{convert_to_singscore}
\title{Convert data sets to run under the method of interest.}
\usage{
convert_to_(dataset, clean)

convert_to_singscore(
  dataset,
  .source,
  .target,
  .target_profile = NULL,
  minsize
)
}
\arguments{
\item{dataset}{A data frame or data frame extension (e.g. a tibble) to convert.}

\item{clean}{Logical value that indicates whether to keep only the columns
necessary to run the function \code{convert_to_[fun]}.}

\item{.source}{column name of geneset/pathway in dataset}

\item{.target}{dataset column name of target mode of regulation,s sign,etc.}

\item{.target_profile}{dataset column name that indicates
\code{\link[=directed2singscore]{directed2singscore()}}).}

\item{minsize}{Minimum number of genes/targets in the gene set}
}
\value{
Returns a tibble with the necessary columns to evaluate the method
 to which the dataset is being converted.
}
\description{
Convert the data set to the suggested standard for the specified function.
If the default parameters are not modified, then the function sets its own
null values for those columns.
}
\concept{convert_to_ variants}
