% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataset-converters.R
\name{convert_to_}
\alias{convert_to_}
\alias{convert_to_scira}
\alias{convert_to_pscira}
\alias{convert_to_mean}
\alias{convert_to_viper}
\alias{convert_to_gsva}
\title{Convert data sets to run under the method of interest.}
\usage{
convert_to_(dataset, ...)

convert_to_scira(dataset, .source, .target, .mor = NULL)

convert_to_pscira(dataset, .source, .target, .mor = NULL)

convert_to_mean(dataset, .source, .target, .mor = NULL, .likelihood = NULL)

convert_to_viper(dataset, .source, .target, .mor = NULL, .likelihood = NULL)

convert_to_gsva(dataset, .source, .target)
}
\arguments{
\item{dataset}{A data frame or data frame extension (e.g. a tibble) to convert.}

\item{...}{Specific parameters to convert the dataset to the correct format.}

\item{.source}{Column with source nodes.}

\item{.target}{Column with target nodes.}

\item{.mor}{Column with edge mode of regulation (i.e. mor).}

\item{.likelihood}{Column with edge likelihood.}
}
\value{
Returns a tibble with the necessary columns to evaluate the method
 to which the dataset is being converted.
}
\description{
Convert the data set to the suggested standard for the specified function.
If the default parameters are not modified, then the function sets its own
null values for those columns.
}
\concept{convert_to_ variants}
