% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataset-converters.R
\name{convert_to_}
\alias{convert_to_}
\alias{convert_to_scira}
\alias{convert_to_pscira}
\title{Convert data sets to run under the method of interest.}
\usage{
convert_to_(dataset, clean)

convert_to_scira(
  dataset,
  .source,
  .target,
  .target_profile = NULL,
  clean = FALSE
)

convert_to_pscira(
  dataset,
  .source,
  .target,
  .target_profile = NULL,
  clean = FALSE
)
}
\arguments{
\item{dataset}{A data frame or data frame extension (e.g. a tibble) to convert.}

\item{clean}{Logical value that indicates whether to keep only the columns
necessary to run the function \code{convert_to_[fun]}.}

\item{.source}{Column name in regulons with the factors to calculate the enrichment scores.}

\item{.target}{Column name in regulons that relates the labels of the rows in the expression matrix.}

\item{.target_profile}{Column name in regulons that contains the target profile. E.g. MoR.}
}
\value{
Returns a tibble with the necessary columns to evaluate the method
 to which the dataset is being converted.
}
\description{
Convert the data set to the suggested standard for the specified function.
If the default parameters are not modified, then the function sets its own
null values for those columns.
}
\concept{convert_to_ variants}
