% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR-decouple.R
\name{decouple}
\alias{decouple}
\title{decouple}
\usage{
decouple(
  mat,
  network,
  .source,
  .target,
  statistics,
  args = list(NULL),
  show_toy_call = FALSE
)
}
\arguments{
\item{mat}{Matrix to evaluate (e.g. expression matrix).
Target nodes in rows and conditions in columns.
\code{rownames(mat)} must have at least one intersection with the elements
in \code{network} \code{.target} column.}

\item{network}{Tibble or dataframe with edges and it's associated metadata.}

\item{.source}{Column with source nodes.}

\item{.target}{Column with target nodes.}

\item{statistics}{Statistical methods to be coupled.}

\item{args}{A list of argument-lists the same length as \code{statistics} (or length 1).
The default argument, list(NULL), will be recycled to the same length as \code{statistics},
and will call each function with no arguments (apart from \code{mat},
\code{network}, \code{.source} and, \code{.target}).}

\item{show_toy_call}{The call of each statistic must be informed?}
}
\value{
A long format tibble of the enrichment scores for each tf
across the samples. Resulting tibble contains the following columns:
\enumerate{
\item \code{statistic}: Indicates which method is associated with which score.
\item \code{tf}: Source nodes of \code{network}.
\item \code{condition}: Condition representing each column of \code{mat}.
\item \code{score}: Regulatory activity (enrichment score).
\item \code{statistic_time}: Internal execution time indicator.
\item \code{...}: Columns of metadata generated by certain statistics.
}
}
\description{
Calculate the TF activity per sample out of a gene expression matrix by
coupling a regulon network with a variety of statistics.
}
