% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR-package.R
\docType{package}
\name{decoupleR-package}
\alias{decoupleR}
\alias{decoupleR-package}
\title{decoupleR: Package to decouple gene sets from statistics}
\description{
Transcriptome profiling followed by differential gene
    expression analysis often leads to lists of genes that are hard to
    analyze and interpret. Downstream analysis tools can be used to
    summarize deregulation events into a smaller set of biologically
    interpretable features.  In particular, methods that estimate the
    activity of transcription factors (TFs) from gene expression are
    commonly used. It has been shown that the transcriptional targets of a
    TF yield a much more robust estimation of the TF activity than
    observing the expression of the TF itself. Consequently, for the
    estimation of transcription factor activities, a network of
    transcriptional regulation is required in combination with a
    statistical algorithm that summarizes the expression of the target
    genes into a single activity score. Over the years, many different
    regulatory networks and statistical algorithms have been developed,
    mostly in a fixed combination of one network and one algorithm. To
    systematically evaluate both networks and algorithms, we developed
    decoupleR , an R package that allows users to apply efficiently any
    combination provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://saezlab.github.io/decoupleR/}
  \item Report bugs at \url{https://github.com/saezlab/decoupleR/issues}
}

}
\author{
\strong{Maintainer}: Christian H. Holland \email{cholland2408@gmail.com} (\href{https://orcid.org/0000-0002-3060-5786}{ORCID})

Authors:
\itemize{
  \item Jesús Vélez \email{jvelez@lcg.unam.mx} (\href{https://orcid.org/0000-0001-5128-3838}{ORCID})
}

}
\keyword{internal}
