% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR-package.R
\docType{package}
\name{decoupleR-package}
\alias{decoupleR}
\alias{decoupleR-package}
\title{decoupleR: Package to decouple gene sets from statistics}
\description{
Usually functional genomics tools consist of a
    fixed combination of gene sets and statistical approach. To overcome
    this limitation this package allows to combine a variety of gene sets
    with a variety of statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://saezlab.github.io/decoupleR/}
  \item Report bugs at \url{https://github.com/saezlab/decoupleR/issues}
}

}
\author{
\strong{Maintainer}: Christian H. Holland \email{cholland2408@gmail.com} (\href{https://orcid.org/0000-0002-3060-5786}{ORCID})

Other contributors:
\itemize{
  \item Jesús Vélez \email{jvelez@lcg.unam.mx} (\href{https://orcid.org/0000-0001-5128-3838}{ORCID}) [contractor]
}

}
\keyword{internal}
