% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR-package.R
\docType{package}
\name{decoupleR-package}
\alias{decoupleR}
\alias{decoupleR-package}
\title{decoupleR: decoupleR: Ensemble of computational methods to infer biological activities from omics data}
\description{
Many methods allow us to extract biological activities from omics data using information from prior knowledge resources, reducing the dimensionality for increased statistical power and better interpretability. Here, we present decoupleR, a Bioconductor package containing different statistical methods to extract these signatures within a unified framework. decoupleR allows the user to flexibly test any method with any resource. It incorporates methods that take into account the sign and weight of network interactions. decoupleR can be used with any omic, as long as its features can be linked to a biological process based on prior knowledge. For example, in transcriptomics gene sets regulated by a transcription factor, or in phospho-proteomics phosphosites that are targeted by a kinase.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://saezlab.github.io/decoupleR/}
  \item Report bugs at \url{https://github.com/saezlab/decoupleR/issues}
}

}
\author{
\strong{Maintainer}: Pau Badia-i-Mompel \email{pau.badia@uni-heidelberg.de} (\href{https://orcid.org/0000-0002-1004-3923}{ORCID})

Authors:
\itemize{
  \item Jesús Vélez-Santiago \email{jvelezmagic@gmail.com} (\href{https://orcid.org/0000-0001-5128-3838}{ORCID})
  \item Jana Braunger \email{jana.bc@gmx.de} (\href{https://orcid.org/0000-0003-0820-9987}{ORCID})
  \item Celina Geiss \email{celina.geiss@stud.uni-heidelberg.de} (\href{https://orcid.org/0000-0002-8740-706X}{ORCID})
  \item Daniel Dimitrov \email{daniel.dimitrov@uni-heidelberg.de} (\href{https://orcid.org/0000-0002-5197-2112}{ORCID})
  \item Sophia Müller-Dott \email{sophia.mueller-dott@uni-heidelberg.de} (\href{https://orcid.org/0000-0002-9710-1865}{ORCID})
  \item Petr Taus \email{petr.taus@ceitec.muni.cz} (\href{https://orcid.org/0000-0003-3764-9033}{ORCID})
  \item Aurélien Dugourd \email{aurelien.dugourd@bioquant.uni-heidelberg.de} (\href{https://orcid.org/0000-0002-0714-028X}{ORCID})
  \item Christian H. Holland \email{cholland2408@gmail.com} (\href{https://orcid.org/0000-0002-3060-5786}{ORCID})
  \item Ricardo O. Ramirez Flores \email{roramirezf@uni-heidelberg.de } (\href{https://orcid.org/0000-0003-0087-371X}{ORCID})
  \item Julio Saez-Rodriguez \email{pub.saez@uni-heidelberg.de} (\href{https://orcid.org/0000-0002-8552-8976}{ORCID})
}

}
\keyword{internal}
