% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper.R
\name{dorothea2viper}
\alias{dorothea2viper}
\title{Helper function}
\usage{
dorothea2viper(genesets)
}
\description{
Helper function to convert DoRothEA gene sets to standardized gene sets
suitable for the \code{\link[=viper]{viper::viper()}} function.
}
