% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-decoupler-formats.R
\name{.decoupler_mat_format}
\alias{.decoupler_mat_format}
\alias{mat_format}
\title{DecoupleR mat format}
\arguments{
\item{mat}{Matrix to evaluate (e.g. expression matrix).
Target nodes in rows and conditions in columns.
\code{rownames(mat)} must have at least one intersection with the elements
in \code{network} \code{.target} column.}
}
\description{
DecoupleR mat format
}
\seealso{
Other decoupleR formats: 
\code{\link{.decoupler_network_format}}
}
\concept{decoupleR formats}
\keyword{internal}
