% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench-run-prep.R
\name{filter_sets}
\alias{filter_sets}
\title{Helper Function to filter and format the gene set resource}
\usage{
filter_sets(set_source, source_col, filter_col, filter_crit, .minsize, silent)
}
\value{
returns a
}
\description{
Helper Function to filter and format the gene set resource
}
