% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench-summ.R
\name{get_bench_summary}
\alias{get_bench_summary}
\title{Function that provides summary and plots for the benchmark run}
\usage{
get_bench_summary(.res_tibble)
}
\arguments{
\item{.res_tibble}{formatted bench result tibble with roc and prc columns
roc column: Reciever Operator Curve Results (calculated with yardstick)
prc column: Precision-Recall Curve Results (calculated with yardstick)}
}
\value{
A summary list with TF coverage, ROC, AUROC, PRAUC, Run time,
ROC plots, and Heatmap plots
}
\description{
Function that provides summary and plots for the benchmark run
}
