% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-omnipath.R
\name{get_ksn_omnipath}
\alias{get_ksn_omnipath}
\title{OmniPath kinase-substrate network}
\usage{
get_ksn_omnipath(...)
}
\arguments{
\item{...}{Passed to \code{OmnipathR::import_omnipath_enzsub}.}
}
\description{
Retrieve a ready to use, curated kinase-substrate Network from the OmniPath
database.
}
\details{
Import enzyme-PTM network from OmniPath, then filter out anything that is not
phospho or dephosphorilation. Then format the columns for use with decoupleR
functions.
}
