% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-omnipath.R
\name{get_KSN_omnipath}
\alias{get_KSN_omnipath}
\title{Omnipath KSN}
\usage{
get_KSN_omnipath(resources = NULL)
}
\arguments{
\item{ressources}{vector of character with ressource names if specified.}
}
\value{
A \code{network} dataframe with 3 columns:
2. \code{source}: Source nodes of \code{network}.
3. \code{target}: Targets nodes of \code{network}.
4. \code{mor}: mode of regulation, indicates phosphorilations
and dephsphorilations
}
\description{
Generate a ready to use, curated Kinase/Substrate Network from omnipath
}
\details{
Import PTM network from omnipath, then filter out anything that is not
phospho or dephosphorilation. Then format the columns for use with decoupleR
functions.
}
\examples{
KSN_omnipath <- get_KSN_omnipath()

}
