% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-randomize-matrix.R
\name{randomize_matrix}
\alias{randomize_matrix}
\title{Randomize matrix}
\usage{
randomize_matrix(mat, randomize_type = c("rows", "cols_independently"))
}
\arguments{
\item{mat}{Matrix to randomize.}

\item{randomize_type}{How to randomize.}
}
\value{
Randomized matrix
}
\description{
Utility function used in functions that require permutations of the
expression matrix
}
\examples{
mat <- matrix(seq_len(9), ncol = 3)
mat

set.seed(42)
randomize_matrix(mat, randomize_type = "rows")

set.seed(42)
randomize_matrix(mat, randomize_type = "cols_independently")
}
