% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataset-converters.R
\name{rename_net}
\alias{rename_net}
\title{Rename network}
\usage{
rename_net(
  network,
  .source,
  .target,
  .mor = NULL,
  .likelihood = NULL,
  def_mor = 1,
  def_lik = 1
)
}
\arguments{
\item{network}{Tibble or dataframe with edges and it's associated metadata.}

\item{.source}{Column with source nodes.}

\item{.target}{Column with target nodes.}

\item{.mor}{Column with edge mode of regulation (i.e. mor).}

\item{.likelihood}{Column with edge likelihood.}

\item{def_mor}{Default value for .mor when not provided.}

\item{def_lik}{Default value for .likelihood when not provided.}
}
\description{
Renames a given network to these column names: .source, .target, .mor,
.likelihood. If .mor or .likelihood are not provided, then the function sets
them to default values.
}
