% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR-benchmark.R
\name{run_benchmark}
\alias{run_benchmark}
\title{Benchmark gene sets with decouple}
\usage{
run_benchmark(.design, .minsize = 10, .form = T, .perform = T, .silent = T)
}
\arguments{
\item{.design}{location of a json file with run design specifications}

\item{.minsize}{regulon/gene set minimum number of targets/members}

\item{.form}{bool whether to format or not}

\item{.perform}{bool whether to calculate roc and performance summary}
}
\value{
An S4 object of class BenchResult
}
\description{
Benchmark gene sets with decouple
}
