% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR-benchmark.R
\name{run_benchmark}
\alias{run_benchmark}
\title{Benchmark pipeline built on the statistical method wrapper decouple.}
\usage{
run_benchmark(
  .design,
  .minsize = 10,
  .form = TRUE,
  .perform = TRUE,
  .silent = TRUE,
  .downsample_pr = FALSE,
  .downsample_roc = FALSE,
  .downsample_times = 100,
  url_bool = FALSE
)
}
\arguments{
\item{.design}{input tibble used to provide the experimental design for each
benchmark run}

\item{.minsize}{regulon/gene set minimum number of targets/members}

\item{.form}{bool whether to format or not}

\item{.perform}{bool whether to calculate roc and performance summary}

\item{.silent}{bool whether to silence messages and warnings}

\item{.downsample_pr}{whether to downsample precision recall curve TNs}

\item{.downsample_roc}{whether to downsample ROC true negatives}

\item{.downsample_times}{downsampling iterattions}

\item{url_bool}{whether prerequisites are loaded from an url}
}
\value{
An S4 object of class BenchResult  \link{BenchResult}
}
\description{
Benchmark pipeline built on the statistical method wrapper decouple.
}
\seealso{
See \link{input_tibble} for a description of the params/column
of .design (i.e. input tibble).
}
