% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR-consensus.R
\name{run_consensus}
\alias{run_consensus}
\title{Consensus score between methods}
\usage{
run_consensus(df, include_time = FALSE)
}
\arguments{
\item{df}{\code{decouple} data frame result}

\item{include_time}{Should the time per statistic evaluated be informed?}
}
\value{
Updated tibble with the computed consensus score between methods
}
\description{
Function to generate a consensus score between methods from the
result of the \code{decouple} function.
}
