% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-fgsea.R
\name{run_fgsea}
\alias{run_fgsea}
\title{fgsea wrapper}
\usage{
run_fgsea(
  mat,
  network,
  .source = .data$source,
  .target = .data$target,
  force_ties = F,
  options = list(),
  seed = 42
)
}
\arguments{
\item{mat}{Matrix to evaluate (e.g. expression matrix).
Target nodes in rows and conditions in columns.
\code{rownames(mat)} must have at least one intersection with the elements
in \code{network} \code{.target} column.}

\item{network}{Tibble or dataframe with edges and it's associated metadata.}

\item{.source}{Column with source nodes.}

\item{.target}{Column with target nodes.}

\item{options}{A list of named options to pass to
\code{\link[=fgsea]{fgsea::fgsea()}}..
These options should not \code{include}, \code{pathways} or \code{stats}.}

\item{seed}{A single value, interpreted as an integer, or NULL.}
}
\value{
A long format tibble of the enrichment scores for each source
across the conditions. Resulting tibble contains the following columns:
\enumerate{
\item{\code{statistic}}: {Indicates which method is associated with which score.}
\item{\code{source}}: {Source nodes of \code{network}.}
\item{\code{condition}}: {Condition representing each column of \code{mat}.}
\item{\code{score}}: {Regulatory activity (enrichment score).}
\item{\code{...}}: {Columns of metadata generated by fgsea method.}
}
}
\description{
This function is a convenient wrapper for the
\code{\link[=fgsea]{fgsea::fgsea()}} function.
}
