% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-run_gsva.R
\name{run_gsva}
\alias{run_gsva}
\title{GSVA wrapper}
\usage{
run_gsva(mat, network, .source, .target, options = list())
}
\arguments{
\item{mat}{Matrix to evaluate (e.g. expression matrix).
Target nodes in rows and conditions in columns.
\code{rownames(mat)} must have at least one intersection with the elements
in \code{network} \code{.target} column.}

\item{network}{Tibble or dataframe with edges and it's associated metadata.}

\item{.source}{Column with source nodes.}

\item{.target}{Column with target nodes.}

\item{options}{A list of named options to pass to
\code{\link[=gsva]{GSVA::gsva()}}..
These options should not \code{include}, \code{expr} or \code{gset.idx.list}.}
}
\value{
A long format tibble of the enrichment scores for each tf
 across the conditions. Resulting tibble contains the following columns:
 \enumerate{
   \item{\code{statistic}}: {Indicates which method is associated with which score.}
   \item{\code{tf}}: {Source nodes of \code{network}.}
   \item{\code{condition}}: {Condition representing each column of \code{mat}.}
   \item{\code{score}}: {Regulatory activity (enrichment score).}
 }
}
\description{
This function is a convenient wrapper for the
\code{\link[=gsva]{GSVA::gsva()}} function.
}
