% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-mean.R
\name{run_mean}
\alias{run_mean}
\title{SAFE mean}
\usage{
run_mean(
  mat,
  network,
  .source = .data$tf,
  .target = .data$target,
  .mor = .data$mor,
  .likelihood = .data$likelihood,
  minsize = 1,
  times = 2,
  seed = 42,
  sparse = TRUE,
  randomize_type = "rows"
)
}
\arguments{
\item{mat}{Matrix to evaluate (e.g. expression matrix).
Target nodes in rows and conditions in columns.
\code{rownames(mat)} must have at least one intersection with the elements
in \code{network} \code{.target} column.}

\item{network}{Tibble or dataframe with edges and it's associated metadata.}

\item{.source}{Column with source nodes.}

\item{.target}{Column with target nodes.}

\item{.mor}{Column with edge mode of regulation (i.e. mor).}

\item{.likelihood}{Column with edge likelihood.}

\item{minsize}{How many output edges a source node must have to be included
in the analysis?}

\item{times}{How many permutations to do?}

\item{seed}{A single value, interpreted as an integer, or NULL for random number generation.}

\item{sparse}{Should the matrices used for the calculation be sparse?}

\item{randomize_type}{How to randomize the expression matrix.}
}
\value{
A long format tibble of the enrichment scores for each tf
 across the conditions. Resulting tibble contains the following columns:
 \enumerate{
   \item{\code{statistic}}: {Indicates which method is associated with which score.}
   \item{\code{tf}}: {Source nodes of \code{network}.}
   \item{\code{condition}}: {Condition representing each column of \code{mat}.}
   \item{\code{score}}: {Regulatory activity (enrichment score).}
   \item{\code{p_value}}: {p-value for the score of mean method.}
 }
}
\description{
Calculate the activity of a regulon through the conditions in the \code{mat}
matrix by calculating the mean over the expression of all genes.
}
