% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scira.R
\name{run_scira}
\alias{run_scira}
\title{SCIRA (Single Cell Inference of Regulatory Activity)}
\usage{
run_scira(
  mat,
  network,
  .source = .data$tf,
  .target = .data$target,
  .target_profile = .data$mor,
  .sparse = FALSE
)
}
\arguments{
\item{mat}{A named expression matrix. E.g Genes in rows and samples in columns.}

\item{network}{Tibble or dataframe with edges and associated metadata.}

\item{.source}{Column with source nodes.}

\item{.target}{Column with target nodes.}

\item{.target_profile}{Column name in network that contains the target profile. E.g. MoR.}

\item{.sparse}{Logical value indicating if the generated profile matrix should be sparse.}
}
\value{
A long format tibble of the enrichment results for each set of genes
 across the samples. Resulting tibble contains the following columns:
 \enumerate{
   \item \code{source} Resources from the \code{.source} column of the \code{network} data frame.
   \item \code{condition} Samples representing each column of \code{mat}.
   \item \code{score} Regulatory activity of each resource for each condition.
 }
}
\description{
Calculates TF activity according to \href{https://www.biorxiv.org/content/10.1101/553040v1.full.pdf}{SCIRA}.
}
\details{
Estimation of regulatory activity: A linear regression of the expression profile
is performed against the "target profile" of the given TF, where in the target
profile, any member of the regulon will be assigned a +1 for activated actions,
a -1 for inhibitory activations and, finally, a 0 for all genes that are not
a member of the TF regulon. TF activity is then defined as the t-statistic of
this linear regression.
}
