% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scira.R
\name{run_scira}
\alias{run_scira}
\title{SCIRA (Single Cell Inference of Regulatory Activity)}
\usage{
run_scira(
  emat,
  regulons,
  .source = .data$tf,
  .target = .data$target,
  .profile = .data$mor,
  .sparse = FALSE
)
}
\arguments{
\item{emat}{A named expression matrix. E.g Genes in rows and samples in columns.}

\item{regulons}{A data frame of regulons in table format.}

\item{.source}{Column name in regulons with the factors to calculate the enrichment scores.}

\item{.target}{Column name in regulons that relates the labels of the rows in the expression matrix.}

\item{.profile}{Column name in regulons that contains the target profile.}

\item{.sparse}{Logical value indicating if the generated profile matrix should be sparse.}
}
\value{
A long format tibble of the enrichment results for each set of genes
 across the samples. Resulting tibble contains the following columns:
 \enumerate{
   \item \code{source} Resources from the \code{.source} column of the \code{regulons} data frame.
   \item \code{condition} Samples representing each column of \code{emat}.
   \item \code{score} Regulatory activity of each resource for each condition.
 }
}
\description{
Calculates TF activity according to \href{https://www.biorxiv.org/content/10.1101/553040v1.full.pdf}{SCIRA}.
}
\details{
Estimation of regulatory activity: A linear regression of the expression profile
is performed against the "target profile" of the given TF, where in the target
profile, any member of the regulon will be assigned a +1 for activated actions,
a -1 for inhibitory activations and, finally, a 0 for all genes that are not
a member of the TF regulon. TF activity is then defined as the t-statistic of
this linear regression.
}
