% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-scira.R
\name{run_scira}
\alias{run_scira}
\title{SCIRA (Single Cell Inference of Regulatory Activity)}
\usage{
run_scira(
  mat,
  network,
  .source = .data$tf,
  .target = .data$target,
  .mor = .data$mor,
  sparse = FALSE,
  fast = TRUE
)
}
\arguments{
\item{mat}{Matrix to evaluate (e.g. expression matrix).
Target nodes in rows and conditions in columns.
\code{rownames(mat)} must have at least one intersection with the elements
in \code{network} \code{.target} column.}

\item{network}{Tibble or dataframe with edges and it's associated metadata.}

\item{.source}{Column with source nodes.}

\item{.target}{Column with target nodes.}

\item{.mor}{Column with edge mode of regulation (i.e. mor).}

\item{sparse}{Logical value indicating if the generated profile matrix should be sparse.}

\item{fast}{Logical value indicating if the lineal model must be calculated
with \code{\link[speedglm:speedlm.fit]{speedlm.fit()}} or with base
\code{\link[stats:lm]{lm()}}.}
}
\value{
A long format tibble of the enrichment scores for each tf
 across the conditions. Resulting tibble contains the following columns:
 \enumerate{
   \item{\code{statistic}}: {Indicates which method is associated with which score.}
   \item{\code{tf}}: {Source nodes of \code{network}.}
   \item{\code{condition}}: {Condition representing each column of \code{mat}.}
   \item{\code{score}}: {Regulatory activity (enrichment score).}
 }
}
\description{
Calculates TF activity according to \href{https://www.biorxiv.org/content/10.1101/553040v1.full.pdf}{SCIRA}.
}
\details{
Estimation of regulatory activity: A linear regression of the expression profile
is performed against the "target profile" of the given TF, where in the target
profile, any member of the regulon will be assigned a +1 for activated actions,
a -1 for inhibitory activations and, finally, a 0 for all genes that are not
a member of the TF regulon. TF activity is then defined as the t-statistic of
this linear regression.
}
