% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signscore.R
\name{run_singscore}
\alias{run_singscore}
\title{SingScore Wrapper}
\usage{
run_singscore(
  emat,
  tiesMethod,
  dataset,
  .source,
  .target,
  .target_profile = NULL,
  minsize,
  perm_num,
  ncores,
  directed,
  tidy
)
}
\arguments{
\item{emat}{Gene expression matrix}

\item{tiesMethod}{Methods to use when ranking ties: min, max, average}

\item{dataset}{Gene set resource}

\item{.source}{column name of geneset/pathway in dataset}

\item{.target}{dataset column name of target mode of regulation,s sign,etc.}

\item{.target_profile}{dataset column name that indicates
\code{\link[=directed2singscore]{directed2singscore()}}).}

\item{minsize}{Minimum number of genes/targets in the gene set}

\item{perm_num}{Number of permutations to be performed}

\item{ncores}{Number of cores to be used for the calculation}

\item{directed}{Bool: Use to indicated whether the Data set resource counts
directed genesets/pathways or not}

\item{tidy}{Bool: Return in tidy format or not}
}
\value{
A dataframe with p-values for each sample or same thing but as a
cartesian product for gene set groupings (i.e. tidy format)
}
\description{
SingScore Wrapper
}
\details{
Please note that a MultiScore method is exists within singscore
which could be used to produce a DF with scores from a GeneSetCollection,
but this is not feasible because generateNull still has to be applied to
each geneset. Also, all genesets must be directed for the GeneSetCollection
to work with all of them or it would again require a split in un/directed gs.
Also, in tidy format it returns confidence for each gene set, p-values are
not calculated seperately for each confidence level (e.g. if the same gs has
low and high they will have the same p-value - check if appropriate)
}
