% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-viper.R
\name{run_viper}
\alias{run_viper}
\title{VIPER wrapper}
\usage{
run_viper(
  mat,
  network,
  .source = .data$tf,
  .target = .data$target,
  .mor = .data$mor,
  .likelihood = .data$likelihood,
  options = list()
)
}
\arguments{
\item{mat}{Matrix to evaluate (e.g. expression matrix).
Target nodes in rows and conditions in columns.
\code{rownames(mat)} must have at least one intersection with the elements
in \code{network} \code{.target} column.}

\item{network}{Tibble or dataframe with edges and it's associated metadata.}

\item{.source}{Column with source nodes.}

\item{.target}{Column with target nodes.}

\item{.mor}{Column with edge mode of regulation (i.e. mor).}

\item{.likelihood}{Column with edge likelihood.}

\item{options}{A list of named options to pass to
\code{\link[=viper]{viper::viper()}} such as \code{minsize} or \code{method}.
These options should not \code{include}, \code{eset} or \code{regulon}.}
}
\value{
A long format tibble of the enrichment scores for each tf
across the conditions. Resulting tibble contains the following columns:
\enumerate{
\item \code{statistic}: Indicates which method is associated with which score.
\item \code{tf}: Source nodes of \code{network}.
\item \code{condition}: Condition representing each column of \code{mat}.
\item \code{score}: Regulatory activity (enrichment score).
\item \code{statistic_time}: Internal execution time indicator.
}
}
\description{
This function is a convenient wrapper for the
\code{\link[=viper]{viper::viper()}} function.
}
