% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-viper.R
\name{run_viper}
\alias{run_viper}
\title{VIPER wrapper}
\usage{
run_viper(emat, genesets, options = list(), gs_resource, tidy = FALSE)
}
\arguments{
\item{emat}{An expression matrix with genes (HGNC symbol) in rows and samples
in columns.}

\item{genesets}{A data frame of gene sets. The structure is dependent on the
gene set resource.}

\item{options}{A list of named options to pass to
\code{\link[=viper]{viper::viper()}} such as \code{minsize} or \code{method}.
 These options should not \code{include}, \code{eset} or \code{regulon}.}

\item{gs_resource}{A character indicating the gene set resource. Based on
this argument \code{genesets} is processed futher via diverse helper
functions (e.g. \code{\link[=dorothea2viper]{dorothea2viper()}}).}

\item{tidy}{Logical, whether computed viper scores should be returned in a
tidy format.}
}
\value{
A matrix of normalized enrichment scores for each gene set across all
 samples. If \code{tidy} is TRUE the normalized enrichment scores are retured
 in a tidy format.
}
\description{
This function is a convenient wrapper for the
\code{\link[=viper]{viper::viper()}} function.
}
