% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signscore.R
\name{simpleScore_undir}
\alias{simpleScore_undir}
\title{Function to calculate singScore and p-values for undirected sets
which implements the \code{\link[=singscore]{singscore::simpleScore()}} and
implements the \code{\link[=singscore]{singscore::getPvals()}} functions.}
\usage{
simpleScore_undir(
  gs_resource,
  geneset,
  rankData,
  perm_num,
  ncores,
  directed,
  permutate
)
}
\arguments{
\item{gs_resource}{dorothea regulon gene sets}

\item{rankData}{ranked expr matrix}

\item{perm_num}{number of permutations to be performed}

\item{ncores}{number of cores to be used}

\item{permutate}{Bool: TRUE if we want to permute and obtain permutate,
FALSE if we are interested only in the scores of the dataset}
}
\value{
A list with 3 data frames (if permutate is TRUE):
scoredf - data.frame of singscores and dispersions for each sample
(only scoredf is returned if permutate is FALSE)
permuteResult - a matrix of scored randomly generated gene sets with the same
number of genes as the gene set
pvals - vector with p-values per expr matrix sample for the given gene set
}
\description{
Function to calculate singScore and p-values for undirected sets
which implements the \code{\link[=singscore]{singscore::simpleScore()}} and
implements the \code{\link[=singscore]{singscore::getPvals()}} functions.
}
\keyword{internal}
