% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signscore.R
\name{simpleScore_undir}
\alias{simpleScore_undir}
\title{Function to calculate singScore and p-values for undirected sets
which implements the \code{\link[=singscore]{singscore::simpleScore()}} and
implements the \code{\link[=singscore]{singscore::getPvals()}} functions.}
\usage{
simpleScore_undir(gs_resource, geneset, rankData, perm_num, ncores, directed)
}
\arguments{
\item{gs_resource}{dorothea regulon gene sets}

\item{rankData}{ranked expression matrix}

\item{perm_num}{number of permutations to be performed}

\item{ncores}{number of cores to be used}

\item{directed}{}
}
\value{
A matrix of p-values
}
\description{
Function to calculate singScore and p-values for undirected sets
which implements the \code{\link[=singscore]{singscore::simpleScore()}} and
implements the \code{\link[=singscore]{singscore::getPvals()}} functions.
}
\keyword{internal}
