% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signscore.R
\name{undirected2singscore}
\alias{undirected2singscore}
\title{Helper function}
\usage{
undirected2singscore(genesets, minsize)
}
\arguments{
\item{genesets}{Undirected genesets genesets}

\item{minsize}{Minimum number of genes regulated}
}
\value{
a named list of lists where each geneset is itself a list with the
genes/targets within the geneset
}
\description{
Helper function to convert undirected resources (e.g. RegNetwork)
to singscore format suitable for the
\code{\link[=singscore]{singscore::simpleScore()}} function.
}
\keyword{internal}
